+++ tests/test_blobs.py
@@ -13,7 +13,8 @@ def prior():
 
 @pytest.mark.parametrize("dtype", [np.float64, np.int64])
 @pytest.mark.parametrize("vectorized", [True, False])
-def test_blobs_single(prior, dtype, vectorized):
+@pytest.mark.parametrize("discard_exploration", [True, False])
+def test_blobs_single(prior, dtype, vectorized, discard_exploration):
     # Test that blobs work and their data types and values match.
 
     def likelihood(x):
@@ -24,15 +25,17 @@ def test_blobs_single(prior, dtype, vectorized):
 
     sampler = Sampler(prior, likelihood, n_dim=2, n_live=10,
                       vectorized=vectorized)
-    sampler.run(f_live=1.0, n_eff=0)
+    sampler.run(f_live=1.0, n_eff=200, discard_exploration=discard_exploration)
     points, log_w, log_l, blobs = sampler.posterior(return_blobs=True)
 
+    assert len(points) == len(blobs)
     assert blobs.dtype == dtype
     assert np.all((10 * points[:, 0]).astype(dtype) == blobs)
 
 
 @pytest.mark.parametrize("vectorized", [True, False])
-def test_blobs_multi(prior, vectorized):
+@pytest.mark.parametrize("discard_exploration", [True, False])
+def test_blobs_multi(prior, vectorized, discard_exploration):
     # Test that blobs work and their data types and values match.
 
     def likelihood(x):
@@ -44,9 +47,10 @@ def test_blobs_multi(prior, vectorized):
 
     sampler = Sampler(prior, likelihood, n_dim=2, n_live=10,
                       vectorized=vectorized)
-    sampler.run(f_live=1.0, n_eff=0)
+    sampler.run(f_live=1.0, n_eff=200, discard_exploration=discard_exploration)
     points, log_w, log_l, blobs = sampler.posterior(return_blobs=True)
 
+    assert len(points) == len(blobs)
     assert blobs['blob_0'].dtype == np.float64
     assert blobs['blob_1'].dtype == np.float32
     assert np.all(points[:, 0].astype(np.float64) == blobs['blob_0'])
@@ -54,7 +58,8 @@ def test_blobs_multi(prior, vectorized):
 
 
 @pytest.mark.parametrize("vectorized", [True, False])
-def test_blobs_dtype(prior, vectorized):
+@pytest.mark.parametrize("discard_exploration", [True, False])
+def test_blobs_dtype(prior, vectorized, discard_exploration):
     # Test that blobs work and their data types and values match.
 
     def likelihood(x):
@@ -66,9 +71,10 @@ def test_blobs_dtype(prior, vectorized):
     blobs_dtype = [('a', '|S10'), ('b', np.int16)]
     sampler = Sampler(prior, likelihood, n_dim=2, n_live=10,
                       vectorized=vectorized, blobs_dtype=blobs_dtype)
-    sampler.run(f_live=1.0, n_eff=0)
+    sampler.run(f_live=1.0, n_eff=200, discard_exploration=discard_exploration)
     points, log_w, log_l, blobs = sampler.posterior(return_blobs=True)
 
+    assert len(points) == len(blobs)
     assert blobs['a'].dtype == blobs_dtype[0][1]
     assert blobs['b'].dtype == blobs_dtype[1][1]
     assert np.all(points[:, 0].astype(blobs_dtype[0][1]) == blobs['a'])